requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
beautifulsoup4==4.8.2
bert-serving-client==1.10.0
bert-serving-server==1.10.0
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
decorator==4.4.1
en-core-web-lg==2.2.5
en-core-web-sm==2.2.5
gast==0.3.3
GPUtil==1.4.0
grpcio==1.26.0
h5py==2.10.0
html2markdown==0.1.7
html2text==2020.1.16
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
joblib==0.14.1
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.1
pandas==1.0.0
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.2
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
soupsieve==1.9.5
spacy==2.2.3
srsly==1.0.1
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
thinc==7.3.1
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.8
Werkzeug==0.16.0
zipp==2.1.0
